requirements.txt
streamlit==1.38.0
requests==2.32.3
python-dotenv==1.0.1
transformers==4.44.2
torch==2.4.1